Enable -Wconversion for thread.cc
This should help prevent bugs due to unexpected implicit integer
conversions.

Bug: 165843530

+++ libartbase/base/utils.h
@@ -40,7 +40,7 @@ static inline uint32_t PointerToLowMemUInt32(const void* p) {
 }
 
 // Returns a human-readable size string such as "1MB".
-std::string PrettySize(int64_t size_in_bytes);
+std::string PrettySize(size_t size_in_bytes);
 
 // Splits a string using the given separator character into a vector of
 // strings. Empty strings will be omitted.
